JSON and ActiveRecord 1) added the score related methods to a module Score in models 2) create migration for pg JSON dictionary data
