+++ app/views/layouts/application.html.erb
@@ -18,6 +18,9 @@
     <div>
       Click <%= link_to "here", user_path(current_user.id) %> to see all your reviews and ratings
     </div>
+    <div>
+      Click <%= link_to "here", '#' %> to see the top 10 users with similar tastes as you.
+    </div>
   <% end %>
 </div>
 <hr>
