+++ db/migrate/20141007172117_create_dictionaries.rb
@@ -1,0 +1,9 @@
+class CreateDictionaries < ActiveRecord::Migration
+  def change
+    create_table :dictionaries do |t|
+      t.json :payload
+
+      t.timestamps
+    end
+  end
+end
