+++ db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20141006131302) do
+ActiveRecord::Schema.define(version: 20141007172117) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -33,6 +33,12 @@ ActiveRecord::Schema.define(version: 20141006131302) do
     t.string   "name"
     t.string   "location"
     t.integer  "brewer_id"
+    t.datetime "created_at"
+    t.datetime "updated_at"
+  end
+
+  create_table "dictionaries", force: true do |t|
+    t.json     "payload"
     t.datetime "created_at"
     t.datetime "updated_at"
   end
